findWarehouseByCity, addClient, updateClient

+++ src/main/java/com/jsp/whs/requestdto/StorageRequest.java
@@ -16,7 +16,7 @@ public class StorageRequest {
 	
 	private String blockName;
 	private String section;
-	private double capacityInWeight;
+	private int capacityInWeight;
 	private double lengthInMeters;
 	private double breadthInMeters;
 	private double heightInMeters;
